CVI_LV_Input: process method now adjusts timetag with value     echoed back from LabView.
Changed some size_t's in Sample, SampleTag and Variable to unsigned ints.
    Want same behavior on 64 and 32 bit systems.

Updates to Resamplers and StatisticsCrunchers so that variable
    indexing is done with unsigned ints.

+++ src/nidas/core/Variable.cc
@@ -224,7 +224,7 @@ void Variable::fromDOMElement(const xercesc::DOMElement* node)
 		setUnits(aval);
 	    else if (aname == "length") {
 	        istringstream ist(aval);
-		size_t val;
+		unsigned int val;
 		ist >> val;
 		if (ist.fail())
 		    throw n_u::InvalidParameterException(
